A6020: Move hardware specific .rc files to /vendor/etc/init/hw
The reason these files are moved to /vendor/etc/init/hw and
not /vendor/etc/init is because 'init' scans the /vendor/etc/init folder
for vendor service .rc fragments to import. That would either result
in duplicate imports or out-of-order imports of these hardware
specific .rc files and that *may* have unintended consequences.

In order to preserve the current import order for MTP, we move all
hardware specific .rc files to /vendor/etc/init/hw which is
autoscanned by init.

Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>